Moved fire scan function into own file in the functions directory
